Provided error handling for setting a sticker to "none"
Setting a sticker to none previously corrupted the data for a sticker
pack resource, making it unusable. The error came about because the
code was trying to load a texture from the project directory, but it
could not find a file when the texture was set to none. Error handling
now appropriately warns the user that they have a null image, but does
not corrupt the data.

Export settings have been updated as well to not allow null stickers
into the xcode build.